\docType{data}
\name{gene2transcript}
\alias{gene2transcript}
\title{Small data frame specifying a transcript to certain genes for synonymous use.}
\format{data frame
\describe{
 \item{gene_name}{HGNC gene name}
 \item{gene_ID}{Ensembl gene ID}
 \item{transcript_ID}{Ensembl transcript ID}
}}
\usage{
gene2transcript
}
\description{
Small data frame specifying a transcript to certain genes for synonymous use.
}
\examples{
\donttest{
 gene2transcript
}
}
\keyword{datasets}


