\docType{data}
\name{referenceDnaStringSet}
\alias{referenceDnaStringSet}
\title{Small DNAStringset as exemplary reference genome sequence}
\format{DNAStringset
\describe{
 \item{width}{Length of feature sequence}
 \item{seq}{Sequence of the feature}
 \item{names}{Name of the feature}
}}
\usage{
referenceDnaStringSet
}
\description{
Small DNAStringset as exemplary reference genome sequence 
}
\examples{
\donttest{
referenceDnaStringSet
}
}
\keyword{datasets}


